download_data.sh now also downloads ltr feature extraction test data.

+++ download_data.sh
@@ -18,8 +18,9 @@ CHECKSUM_FILE="sha256sums"
 download()
 {
   cd "$CALLING_BASE_DIR"
-  NAME="$1"
-  VERSION="$2"
+  TREAT_AS="$1"
+  NAME="$2"
+  VERSION="$3"
   ARCHIVE_BASENAME="${NAME}_$VERSION"
   ARCHIVE_NAME="$ARCHIVE_BASENAME.tgz"
   URL="http://s3-de-central.profitbricks.com/xayn-yellow-bert/$NAME/$ARCHIVE_NAME"
@@ -33,12 +34,15 @@ download()
   cd "$ARCHIVE_BASENAME"
   shasum -c "$CHECKSUM_FILE"
 
-  # update symlinks (the `|| :` makes the shell ignore failure of rm even if `set -eu` is used)
-  cd "$CALLING_BASE_DIR"
-  rm "${SELF_DIR_PATH}/bindings/dart/example/assets/${NAME}"_v* || :
-  ln -s "../../../../data/${ARCHIVE_BASENAME}" "${SELF_DIR_PATH}/bindings/dart/example/assets/${ARCHIVE_BASENAME}"
+  if [ "$TREAT_AS" = "asset" ]; then
+    # update symlinks (the `|| :` makes the shell ignore failure of rm even if `set -eu` is used)
+    cd "$CALLING_BASE_DIR"
+    rm "${SELF_DIR_PATH}/bindings/dart/example/assets/${NAME}"_v* || :
+    ln -s "../../../../data/${ARCHIVE_BASENAME}" "${SELF_DIR_PATH}/bindings/dart/example/assets/${ARCHIVE_BASENAME}"
+  fi
 }
 
-download smbert v0000
-download qambert v0001
-download ltr v0000
+download asset smbert v0000
+download asset qambert v0001
+download asset ltr v0000
+download misc ltr_feature_extraction_tests v0000
